Processor.getSize() added, some more updates too. Now pretty printing sizes.

+++ src/main/kotlin/ProcessingProcess.kt
@@ -5,9 +5,9 @@ class ProcessingProcess {
     var profile: Profile? = null
     var terminal: Terminal? = null
     var successfulEntries: Int = 0
-    var successfulBytes: Int = 0
+    var successfulBytes: Long = 0
     var failedEntries: Int = 0
-    var failedBytes: Int = 0
+    var failedBytes: Long = 0
 
     override fun toString(): String {
         return "ProcessingProcess: $profile"
